scanner-pubsub/requirements.txt
google-cloud-pubsub>=2.16.0
google-cloud-storage>=3.1.0